Move media data task logic to Client
