+++ Cartfile.resolved
@@ -3,4 +3,3 @@ github "Quick/Nimble" "v5.1.1"
 github "anviking/Decodable" "v0.5.1"
 github "krzyzanowskim/CryptoSwift" "0.6.7"
 github "venmo/DVR" "v1.0.0"
-
