+++ Sources/Asset.swift
@@ -7,11 +7,6 @@
 //
 
 import Foundation
-import Interstellar
-
-#if os(iOS) || os(tvOS)
-import UIKit
-#endif
 
 /// An asset represents a media file in Contentful
 public struct Asset: Resource, LocalizedResource {
@@ -27,8 +22,13 @@ public struct Asset: Resource, LocalizedResource {
 
     /// The unique identifier of this Asset
     public let identifier: String
+
     /// Resource type ("Asset")
     public let type: String
+
+    /// Currently selected locale
+    public var locale: String
+
     /// The URL for the underlying media file
     public func URL() throws -> Foundation.URL {
         if let urlString = (fields["file"] as? [String: Any])?["url"] as? String {
@@ -41,35 +41,4 @@ public struct Asset: Resource, LocalizedResource {
 
         throw SDKError.invalidURL(string: "")
     }
-
-    /// Currently selected locale
-    public var locale: String
-
-    fileprivate let networkingSession = NetworkingSession()
-
-    /**
-     Fetch the underlying media file as `Data`
-
-     - returns: Tuple of the data task and a signal for the `NSData` result
-     */
-    public func fetch() -> (URLSessionDataTask?, Observable<Result<Data>>) {
-        do {
-            return networkingSession.fetch(url: try URL())
-        } catch let error {
-            let signal = Observable<Result<Data>>()
-            signal.update(Result.error(error))
-            return (URLSessionDataTask(), signal)
-        }
-    }
-
-#if os(iOS) || os(tvOS)
-    /**
-     Fetch the underlying media file as `UIImage`
-
-     - returns: Tuple of data task and a signal for the `UIImage` result
-     */
-    public func fetchImage() -> (URLSessionDataTask?, Observable<Result<UIImage>>) {
-        return convert_signal(closure: fetch) { UIImage(data: $0) }
-    }
-#endif
 }
