+++ Sources/SignalUtils.swift
@@ -39,40 +39,3 @@ func signalify<U>(closure: SignalBang<U>) -> (URLSessionDataTask?, Observable<Re
     let value = closure { signal.update($0) }
     return (value, signal)
 }
-
-class NetworkingSession {
-    var sessionConfigurator: ((URLSessionConfiguration) -> Void)?
-
-    fileprivate var urlSession: URLSession {
-        let sessionConfiguration = URLSessionConfiguration.default
-        if let sessionConfigurator = sessionConfigurator {
-            sessionConfigurator(sessionConfiguration)
-        }
-        return URLSession(configuration: sessionConfiguration)
-    }
-
-    func fetch(url: URL, completion: @escaping (Result<Data>) -> Void) -> URLSessionDataTask {
-        let task = urlSession.dataTask(with: url) { data, response, error in
-            if let data = data {
-                completion(Result.success(data))
-                return
-            }
-
-            if let error = error {
-                completion(Result.error(error))
-                return
-            }
-
-            let sdkError = SDKError.invalidHTTPResponse(response: response)
-            completion(Result.error(sdkError))
-        }
-
-        task.resume()
-        return task
-    }
-
-    func fetch(url: URL) -> (URLSessionDataTask?, Observable<Result<Data>>) {
-        let closure: SignalObservation<URL, Data> = fetch
-        return signalify(parameter: url, closure: closure)
-    }
-}
