+++ Tests/AssetTests.swift
@@ -83,6 +83,7 @@ class AssetTests: XCTestCase {
 
         AssetTests.client.fetchAsset(identifier: "nyancat").1.then { asset in
             AssetTests.client.fetchImage(for: asset).1.then { image in
+
                 expect(self.md5(image)).to(equal("94fd9a22b0b6ecab15d91486922b8d7e"))
                 expectation.fulfill()
             }
